+++ 2_Android/MusicBox/base-material/src/main/java/com/chenyuwei/basematerial/activity/BaseActivity.java
@@ -63,8 +63,8 @@ public abstract class BaseActivity extends AppCompatActivity implements View.OnC
         }
     }
 
-    public int getUid(){
-        return preferences.getInt("uid",-1);
+    public long getUid(){
+        return preferences.getLong("uid",-1);
     }
 
     protected void toast(String message) {
